requirements.txt
# API Clients
ccxt>=1.93.0
python-binance>=1.0.16
coinbase>=2.1.0
kucoin-python>=1.0.11

# Data Analysis
pandas>=1.3.5
numpy>=1.21.6
ta>=0.10.1
scikit-learn>=1.0.2
tensorflow>=2.8.0
keras>=2.8.0

# On-Chain Data
# Note: These packages are not available in PyPI, we'll implement custom clients
# glassnode>=1.0.0
# cryptoquant>=0.1.0

# Sentiment Analysis
tweepy>=4.10.0
praw>=7.6.0
newsapi-python>=0.2.6
textblob>=0.17.1
nltk>=3.7

# Database
sqlalchemy==1.4.47  # Pinned version - DO NOT UPGRADE to 2.x without fixing model conflicts
psycopg2-binary>=2.9.3
alembic>=1.8.0

# Web Dashboard
flask>=2.1.2
flask-login>=0.6.1
flask-wtf>=1.0.1
dash>=2.5.1
plotly>=5.8.2

# AWS
boto3>=1.24.13

# Utilities
python-dotenv>=0.20.0
schedule>=1.1.0
loguru>=0.6.0
tqdm>=4.64.0
